Change to *sh -> fish bootstrapping
The first time we run `sh/profile` we re-exec into fish.  When this is
done, the original goal was to set `SHELL` so that any tools deeper down
the process tree would go straight to fish and skip the sh / bash / zsh
steps when starting an interactive shell.  The assumption was these
processes would inherit their parent settings, notably `PATH`.

However, `tmux` does _not_ do this, and instead launchs $SHELL as a
login shell.  In macOS this means it picks up a bunch of random paths
that I don't want.

This patch forces all new shells in tmux to go the sh -> fish bootstrap
route, setting a very specific path.